move prom/ha/configapi to prom/instance/configstore/configapi
